refactor(ui): colocate api-reference components into _components/
First of the colocation follow-ups after the App Router migration closeout.
Establishes the convention for the remaining ~28 segments: a page's owned
components move into `<segment>/_components/` (private), and components shared
across segments move to the shared `@/components` home rather than sitting in
one segment's folder.

For api-reference:
- `APIReferenceView` and `DocLink` are api-reference-only, so they move into
  `api-reference/_components/`
- `CodeBlock` is imported by playground and cost-tracking too, so it is not
  api-reference-owned; it moves to `@/components/CodeBlock` and the three
  external importers are repointed there

No behavior change; pure file relocation plus import updates. The grandfathered
`@tremor/react` suppression on `APIReferenceView` is re-keyed to its new path
(converting that view off tremor is a separate effort, out of scope for a move).
