+++ ui/litellm-dashboard/eslint-suppressions.json
@@ -4,7 +4,7 @@
       "count": 1
     }
   },
-  "src/app/(dashboard)/api-reference/APIReferenceView.tsx": {
+  "src/app/(dashboard)/api-reference/_components/APIReferenceView.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
